pyproject.toml
[project]
name = "wordclouddemo"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "mizuki", email = "trombonelover8888@gmail.com" }
]
dependencies = ["mecab-python3~=1.0.6", "unidic-lite~=1.0.8", "wordcloud~=1.9.2", "matplotlib~=3.7.1"]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
[tool.hatch.metadata]
allow-direct-references = true
